python-app/requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
gunicorn==20.1.0
icecream==2.1.1
pandas==1.3.4
pydantic==1.8.2
python-decouple==3.5
pymongo==3.12.0
scikit_learn==1.0.1
fastapi-utils==0.2.1
scikit-surprise==1.1.1